fix: let the engine refuse what the merge gate would refuse
The gate this PR added holds a partial run back whole when its literal
parity fails, which is right -- the key it would land is a backfill no
later run re-queues, so a dropped identifier that lands is frozen. But
_validate never asked the question the gate asks. It checks emptiness,
hardcoded names, placeholders, markup and panel links, and not one
number or literal, so the engine could still produce the Russian
docs/beta.md drop and the Chinese enable_tool_search localisation this
PR repairs by hand. With the gate in place that output no longer lands
quietly -- it reddens the push instead, discards the progress file, and
sends the next run to re-plan the same work on a fresh quota. A
permanent single-string fault becomes a permanent whole-pipeline stall.

So the engine now runs the same comparison before accepting a string.
One rejection, one retry, and the key is left for the next run to plan
again; the push gate stays as the backstop behind it.

One comparison, one home: the rules move to scripts/locale_rules.py and
both callers import them. Two copies would mean the engine accepting
exactly what the gate later refuses, which is the failure this change
exists to prevent, written twice.

The engine deliberately skips one arm -- the number multiset -- and the
tolerance table is the evidence. Both entries in it are number-count
entries: Russian spells "the 5 experimental sub-flags" out in words,
Chinese keeps a clause the English rendering cuts. Comparing multisets
without a per-key tolerance would refuse correct output on every run and
stall exactly as described above, one step upstream. All three faults
this PR repaired by hand are found by the arms that do run.

Also corrected, all measured against the tree:

- 88 tools and 176 strings, not 87 and 174; 84 of the 88 display their
  summary paragraph verbatim, not 84 of 87.
- "Every shipped locale translates all 174 today" is no longer true of
  this branch: nl, pl and sv carry 176, de, fr and ru 169, and es, it
  and zh-Hans 175.
- "the component version 1.2.4" is a minimum a help string states; the
  component is on 1.3.3.
- Seven tool keys differ between renderings, not eleven.
- The workflow header promised that a partial run always pushes what
  finished. This PR added the one exception, so the header says so.
- Three reasons in the gate-shape test claimed that setting the
  completeness variable on the partial step would drag the gated tests
  in. It would not: the -k filter deselects them and the selected test
  carries no marker, so the variable is inert there. The assertion is
  worth keeping for the pair it forms with the filter, and now says why.
