+++ tests/src/unit/test_locale_parity.py
@@ -56,6 +56,17 @@ _REPO_ROOT = Path(__file__).resolve().parents[3]
 sys.path.insert(0, str(_REPO_ROOT / "scripts"))
 import extract_tools  # noqa: E402
 import generate_locales  # noqa: E402
+from locale_rules import (  # noqa: E402
+    _canonical_number,
+    _invented_files,
+    _localised_hardcoded_name,
+    _lost_literals,
+    _lost_magnitudes,
+    _numbers,
+    _parity_fault,
+    _reversed_ordered_pairs,
+    _untranslatable_names,
+)
 
 SETTINGS_LOCALES = _REPO_ROOT / "src" / "ha_mcp" / "settings_ui" / "locales"
 AGENTS_MD = _REPO_ROOT / "AGENTS.md"
@@ -257,7 +268,7 @@ def _english_tool_sources() -> dict[str, str]:
     """The English tool texts a settings UI catalog translates.
 
     ``en.json`` leaves ``tools`` empty — English for those comes from the tool
-    definitions at runtime — so these 174 strings live in no catalog and the
+    definitions at runtime — so these 176 strings live in no catalog and the
     baseline did not cover them. An edit to a docstring therefore left every
     locale describing the old behaviour with nothing going red:
     ``ha_dev_manage_settings`` gained the Tools/Policies/Backups surfaces and
@@ -273,7 +284,7 @@ def _english_tool_sources() -> dict[str, str]:
     pinned.
 
     What gets hashed is the summary *paragraph*, while the row displays its
-    first physical line cut at 120 characters — the same text for 84 of the 87
+    first physical line cut at 120 characters — the same text for 84 of the 88
     tools. ``ha_config_set_helper`` wraps its summary and the Chinese catalog
     translates the half that wraps off, so hashing the displayed text would
     leave "(28 types, unified interface)" free to move while a shipped
@@ -1052,11 +1063,13 @@ def test_generated_addon_projections_are_translated(locale: str) -> None:
 def test_settings_catalog_tools_are_translated(locale: str) -> None:
     """Exactness says every tool is present, not that any was translated.
 
-    87 titles and 87 descriptions is the largest translated surface in the
+    88 titles and 88 descriptions is the largest translated surface in the
     repo, and nothing looked at the values. The exactness rule above also
     obliges every tool-adding PR to touch six languages before it can go
     green, which is pressure toward pasting the English in — this is what
-    notices. Every shipped locale translates all 174 today.
+    notices. nl, pl and sv carry all 176 today; the values #2180 dropped for
+    the sync to rewrite leave de, fr and ru at 169 and es, it and zh-Hans
+    at 175.
     """
     english = _english_tool_texts()
     catalog = _settings_catalog(locale).get("tools", {})
@@ -1151,112 +1164,6 @@ def test_component_catalog_is_not_a_copy_of_english(locale: str) -> None:
     )
 
 
-# A digit run preceded by a letter or another digit belongs to an identifier
-# ("Z2M", "MQTT5"), not to a claim about quantity, so it is not a number a
-# translation owes. A trailing unit is deliberately left out of the token:
-# a locale that writes "5 тыс." for "5K" still matches on the digits, while
-# "46K" where the English says 90% and 5K still reports as changed.
-#
-# Separators inside a number are kept as GROUP BOUNDARIES rather than deleted.
-# Deleting them tolerates locale punctuation but also erases the difference
-# between the "4.5:1" contrast ratio and "45:1", and between the component
-# version "1.2.4" and "12.4" -- both live English strings. Comparing the tuple
-# of groups keeps "4.5" and "4,5" equal while "45" stays a different number.
-_NUMBER_RE = re.compile(r"(?<![A-Za-z0-9])\d+(?:[.,   ]\d+)*")
-# A magnitude suffix is part of the claim: "5K" and "5M" share their digits.
-_MAGNITUDE_RE = re.compile(r"(?<![A-Za-z0-9])(\d+)([KMGT])(?![A-Za-z])")
-# A percentage is a unit every catalog keeps, spaced or not: four write
-# "90 %" and four "90%", so the sign is required but the space is free.
-_PERCENT_RE = re.compile(r"(?<![A-Za-z0-9])(\d+)\s?%")
-# A storage unit is part of the claim too: "1-256 MB" and "1-256 GB" differ by
-# three orders of magnitude while the digits match.
-_COMPOUND_UNIT_RE = re.compile(
-    r"(?<![A-Za-z0-9])(\d+)\s*([A-Za-zА-Яа-я]{2})(?![A-Za-zА-Яа-я])"
-)
-_KNOWN_UNITS = frozenset({"KB", "MB", "GB", "TB"})
-# The same four units as every catalog spells them. Without this the filter
-# that stops a localised spelling false-positiving also stops a *wrong*
-# localised spelling reporting: "предел 1-256 ГБ" and "limite de 1-256 Go"
-# both state gigabytes where the English says megabytes, and both passed.
-# Russian transliterates the prefix and abbreviates "байт"; French sets "o"
-# for "octet". Only spellings a shipped catalog actually uses are listed --
-# an unknown two-letter token stays uncomparable, which is the safe arm.
-_LOCALISED_UNITS = {
-    "КБ": "KB",
-    "МБ": "MB",
-    "ГБ": "GB",
-    "ТБ": "TB",
-    "Ko": "KB",
-    "Mo": "MB",
-    "Go": "GB",
-    "To": "TB",
-}
-
-
-def _canonical_unit(unit: str) -> str | None:
-    """One storage unit as its English spelling, or None if not comparable."""
-    if unit.upper() in _KNOWN_UNITS:
-        return unit.upper()
-    return _LOCALISED_UNITS.get(unit)
-# "N > 0" and "N < 0" are opposite conditions with identical numbers.
-_COMPARISON_RE = re.compile(r"([A-Za-z_]\w*)\s*([<>]=?)\s*(\d+)")
-# A range is an ordered claim, and reversing it leaves the digits untouched:
-# "Range 1-600" and "Range 600-1" hold the same multiset while the second
-# names a lower bound above its upper one. The endpoints are spelled by
-# reference to the number pattern rather than repeated, so a grouped bound
-# ("1 024") keeps reading as one endpoint here too, and every dash a catalog
-# might set counts -- hyphen through em dash. A ratio is the same ordered
-# claim behind a different separator, and the corpus ships one: the "4.5:1"
-# contrast ratio every catalog carries. Reversing it to "1:4,5" leaves the
-# digits untouched exactly as a reversed range does, so a colon counts here
-# too -- both the ASCII one and the fullwidth colon a CJK catalog may set.
-_ORDERED_PAIR_RE = re.compile(
-    rf"({_NUMBER_RE.pattern})\s*([-‐-―:：])\s*({_NUMBER_RE.pattern})(?![A-Za-z0-9])"
-)
-_GROUP_SEPARATOR_RE = re.compile(r"[.,   ]")
-
-# Tokens a reader has to type, search for, or find on disk. Prose slashes
-# ("read/write") are not paths, hence the requirement that a path start at a
-# separator; a bare word is not an identifier, hence the required underscore.
-# Files come FIRST inside the literal pattern below: with snake_case ahead of
-# them, "tool_policy.json" tokenised as "tool_policy" plus ".json", and neither
-# half is the name a reader has to find. "*" belongs in the stem for
-# "packages/*.yaml". Spelled here rather than inline so the reverse check can
-# ask for this shape alone without a second copy of it to keep in step.
-_FILE_PATTERN = r"[\w.~*/-]*\.(?:yaml|yml|json|py|md|txt)"
-_FILE_RE = re.compile(_FILE_PATTERN)
-
-_LITERAL_RE = re.compile(
-    rf"""(?:
-        {_FILE_PATTERN}                         # files: configuration.yaml
-      | [a-z][a-z0-9]*(?:_[a-z0-9]+)+           # snake_case: enable_tool_search
-      | [A-Z][A-Z0-9]*(?:_[A-Z0-9]+)+           # ALL_CAPS: DISABLED_TOOLS
-      | (?<![\w/<])~?/[\w.*-]+(?:/[\w.*-]+)*    # paths: /api/settings/features
-      | [a-z][a-z0-9+.-]*://\S*                 # any scheme, bare one included
-      | (?<![\w.])[a-z][a-z0-9_]*(?:\.[a-z][a-z0-9_]*)+(?![\w])  # group.set
-      | (?<!\w)[A-Za-z][A-Za-z_]*\d+(?!\w)      # Jinja2, alert2
-        # A repository slug, both halves hyphenated. That shape is what keeps
-        # "read/write", "enable/disable" and "re-add/refresh" out -- prose
-        # pairs a slash without hyphenating both sides. A slug that hyphenates
-        # neither would slip through; none ships today.
-      | (?<![\w/])[a-z0-9]+(?:-[a-z0-9]+)+/[a-z0-9]+(?:-[a-z0-9]+)+(?![\w/])
-      | (?<!\w)[A-Z][a-z0-9]+(?:[A-Z][A-Za-z0-9]*)+(?!\w)  # WebSocket, Z2M
-    )""",
-    re.X,
-)
-
-# An exact argument a caller has to pass: yaml_path='automation'. The word
-# alone is ordinary prose a locale should translate, so only the quoted form
-# counts -- and the parameter name is captured with it, because the value on
-# its own says nothing about which argument it belongs to: "scope='snapshot',
-# action='delete'" and "scope='delete', action='snapshot'" share both values
-# and describe opposite calls.
-_QUOTED_ASSIGNMENT_RE = re.compile(r"(?<![\w])([a-z_]+\s*=\s*'[^']+')")
-
-# Latin abbreviations the dotted-identifier arm would otherwise claim. Prose,
-# not names: no reader types "e.g" into anything.
-_PROSE_ABBREVIATIONS = frozenset({"e.g", "i.e"})
-
 # (locale, surface, key) -> why this pair cannot satisfy the rule. Keep the
 # reason specific enough that a later reader can re-decide it; an exception
 # whose justification is "the check is noisy" belongs in the check instead.
@@ -1298,326 +1205,6 @@ LITERAL_PARITY_EXCEPTIONS: dict[
         "missing because the sentence is right, not because a number was lost.",
     ),
 }
-
-
-def _canonical_number(token: str) -> tuple[str, ...]:
-    """One number as its groups, with thousands grouping folded away.
-
-    A separator followed by groups of exactly three digits is grouping, not a
-    boundary: English ships `10000`, `1440` and `65535`, and a locale writing
-    `10.000` states the same number. Anything else keeps its groups, so the
-    "4.5:1" ratio and the version "1.2.4" stay distinct from "45" and "12.4".
-    A decimal written to exactly three places is the ambiguous case and folds
-    with the thousands; no English string has one.
-    """
-    groups = _GROUP_SEPARATOR_RE.split(token)
-    if (
-        len(groups) > 1
-        and len(groups[0]) <= 3
-        and all(len(group) == 3 for group in groups[1:])
-    ):
-        return ("".join(groups),)
-    return tuple(groups)
-
-
-def _numbers(text: str) -> Counter[tuple[str, ...]]:
-    return Counter(_canonical_number(token) for token in _NUMBER_RE.findall(text))
-
-
-def _lost_magnitudes(english: str, translated: str) -> list[str]:
-    """Magnitude suffixes the translation contradicts rather than drops.
-
-    A unit is part of the claim, and the value comparison cannot see it: "5K"
-    and "5M" carry the same digits, and so do "90%" and a bare "90". The
-    percent sign is required outright because every catalog keeps it -- four
-    write "90 %" and four "90%", so only the space is free. Requiring a
-    magnitude suffix outright would be wrong, though: Polish writes
-    "5 tys." and Russian "5 тыс." for it, and both are correct. So the suffix
-    is only compared when the translation itself puts a letter straight
-    onto those digits -- six of the nine catalogs do, Polish and Russian spell
-    it out, and the Italian value is one this branch deletes for the sync to
-    rewrite.
-
-    A storage unit is compared in the alphabet the catalog writes it in, not
-    only in Latin: "предел 1-256 ГБ" and "limite de 1-256 Go" both state
-    gigabytes where the English says megabytes, and while the filter that
-    keeps a *correct* localised spelling from reporting was in place, so was
-    the blind spot for a wrong one.
-
-    Both suffix arms ask whether ANY occurrence of the digits carries the
-    right unit, rather than settling on the first. The first occurrence is not
-    necessarily the matching one -- "bis zu 5x schneller, etwa 5K Token" is
-    faithful and reported, "jusqu'a 256 Go puis 256 MB" contradicts the
-    English and did not.
-    """
-    contradicted = set()
-    for digits, unit in _COMPOUND_UNIT_RE.findall(english):
-        if unit.upper() not in _KNOWN_UNITS:
-            continue
-        # EVERY occurrence of these digits, not the first one: "jusqu'a 256 Go
-        # puis 256 MB" carries the unit on its second, and settling on the
-        # first reported a correct translation as a contradiction while
-        # missing the storage arm's own case. One occurrence spelling the
-        # unit right is the translation carrying the claim.
-        carried = {
-            canonical
-            for match in re.finditer(
-                rf"(?<![A-Za-z0-9]){re.escape(digits)}\s*([A-Za-zА-Яа-я]{{2}})"
-                r"(?![A-Za-zА-Яа-я])",
-                translated,
-            )
-            if (canonical := _canonical_unit(match.group(1))) is not None
-        }
-        if carried and unit.upper() not in carried:
-            contradicted.add(f"{digits} {unit}")
-    for name, operator, digits in _COMPARISON_RE.findall(english):
-        if not re.search(
-            rf"{re.escape(name)}\s*{re.escape(operator)}\s*{digits}", translated
-        ):
-            contradicted.add(f"{name} {operator} {digits}")
-    for digits in _PERCENT_RE.findall(english):
-        if not re.search(rf"(?<![A-Za-z0-9]){re.escape(digits)}\s?%", translated):
-            contradicted.add(f"{digits}%")
-    for digits, suffix in _MAGNITUDE_RE.findall(english):
-        # Same every-occurrence rule as the storage arm. "bis zu 5x schneller,
-        # etwa 5K Token" is a correct translation whose *first* run of these
-        # digits carries an unrelated letter, and reading only that one turned
-        # it red -- post-merge that blocks the push for the whole run.
-        carried = {
-            match.group(1).upper()
-            for match in re.finditer(
-                rf"(?<![A-Za-z0-9]){re.escape(digits)}([A-Za-z])", translated
-            )
-        }
-        if carried and suffix.upper() not in carried:
-            contradicted.add(f"{digits}{suffix}")
-    return sorted(contradicted)
-
-
-def _reversed_ordered_pairs(english: str, translated: str) -> list[str]:
-    """Ordered number pairs the translation states back to front.
-
-    A range is an ordered claim that the value comparison is blind to: both
-    endpoints of "Range 1-600" are still present in "Range 600-1", so the
-    number multiset balances while the text now names a floor above its
-    ceiling.
-
-    Only an actual inversion reports. A catalog is free to write "von 1 bis
-    600" instead of setting a dash, and its numbers stay guarded by the value
-    comparison either way; demanding the dash form back would fail a correct
-    rendering rather than a wrong one.
-
-    A ratio is the same claim with a colon for a dash: "4.5:1" and "1:4,5"
-    hold the same two numbers, and the second states a contrast threshold no
-    checker would ever set. The separator the English used is carried into
-    the report so a ratio is not named as a range.
-    """
-    translated_bounds = {
-        (_canonical_number(low), _canonical_number(high))
-        for low, _, high in _ORDERED_PAIR_RE.findall(translated)
-    }
-    return sorted(
-        {
-            f"{low}{':' if separator in ':：' else '-'}{high}"
-            for low, separator, high in _ORDERED_PAIR_RE.findall(english)
-            if (bounds := (_canonical_number(low), _canonical_number(high)))
-            and bounds[0] != bounds[1]
-            and bounds[::-1] in translated_bounds
-            and bounds not in translated_bounds
-        }
-    )
-
-
-def _show_numbers(counted: Counter[tuple[str, ...]]) -> list[str]:
-    return sorted(".".join(groups) for groups in counted.elements())
-
-
-def _parity_fault(
-    english: str,
-    translated: str,
-    *,
-    tolerated_losses: Counter[tuple[str, ...]] | None = None,
-    tolerated_additions: Counter[tuple[str, ...]] | None = None,
-) -> str:
-    """Everything one English/translated pair contradicts, or "" if nothing.
-
-    One function rather than an expression inline in the test, because the
-    set of arms it runs is itself a thing that can be broken: dropping any
-    one of them from the sum leaves every shipped pair green -- the corpus is
-    clean, so an arm that stops being called reports nothing and looks
-    identical to an arm that found nothing.
-    ``test_the_comparison_runs_every_arm`` holds each one to a case it must
-    report, which only works if there is a single place they are summed.
-
-    Numbers are compared as multisets in both directions; a tolerance is
-    subtracted by occurrence rather than by value.
-    """
-    losses = tolerated_losses if tolerated_losses is not None else Counter()
-    additions = tolerated_additions if tolerated_additions is not None else Counter()
-    lost_numbers = (_numbers(english) - _numbers(translated)) - losses
-    gained_numbers = (_numbers(translated) - _numbers(english)) - additions
-    lost = (
-        _lost_literals(english, translated)
-        + _lost_magnitudes(english, translated)
-        + _reversed_ordered_pairs(english, translated)
-        + _localised_hardcoded_name(english, translated)
-        + _invented_files(english, translated)
-    )
-    return ", ".join(
-        part
-        for part in (
-            f"numbers lost {_show_numbers(lost_numbers)}" if lost_numbers else "",
-            f"numbers added {_show_numbers(gained_numbers)}" if gained_numbers else "",
-            f"identifiers dropped {lost}" if lost else "",
-        )
-        if part
-    )
-
-
-@cache
-def _untranslatable_names() -> frozenset[str]:
-    """On-screen names our own Python hardcodes, which must stay English.
-
-    Read from the pipeline at runtime instead of copied, so this check cannot
-    drift away from ``translate_locales._untranslatable_name_dropped``, the
-    guard that rejects engine output localising one.
-    """
-    import translate_locales
-
-    return frozenset(translate_locales._hardcoded_ui_names())
-
-
-def _without_sentence_punctuation(literal: str) -> str:
-    """A literal with the prose punctuation it swallowed removed.
-
-    A path and a URI both run to the next space, so "see /api/x." and
-    "at skill://y)" carry the sentence's period or bracket into the token, and
-    requiring it verbatim fails a translation that keeps the literal but ends
-    its sentence differently.
-
-    An ellipsis is left alone: ``/api/webhook/...`` is the one live English
-    literal ending in punctuation, and there the dots are part of what the
-    reader is shown, not the end of a sentence.
-    """
-    if literal.endswith("..."):
-        return literal
-    return literal.rstrip(".,;:!?)]}\"'»”")
-
-
-def _carries(literal: str, translated: str) -> bool:
-    """Whether the translation still contains this literal as a whole token.
-
-    Substring, but not *any* substring: the match may not continue into
-    another identifier character, so "enable_tool_search_old" no longer counts
-    as carrying `enable_tool_search`, nor "configuration.yaml.bak" as carrying
-    `configuration.yaml`. A hyphen or a case change is a different matter --
-    German writes "/data-Volume" and Swedish "/data-volymen", and both carry
-    the original intact. A dotted extension does not count either, so
-    "configuration.yaml.bak" is a different file -- and so is
-    "other.configuration.yaml", so a dot is excluded on the left as well --
-    while a sentence-ending period after the name is not part of the token and
-    still matches.
-    """
-    return (
-        re.search(
-            r"(?<![A-Za-z0-9_.])"
-            + re.escape(literal)
-            + r"(?![A-Za-z0-9_]|\.[A-Za-z0-9])",
-            translated,
-        )
-        is not None
-    )
-
-
-def _lost_literals(english: str, translated: str) -> list[str]:
-    """English literals absent from the translation, as substrings.
-
-    Substring rather than token equality on purpose: German writes
-    "/data-Volume" and Swedish "/data-volymen" and "skill://-resurs", which
-    tokenise as different literals while carrying the original intact.
-    Re-extracting from the translation reports every one of them as a loss;
-    asking whether the English literal is still findable reports none.
-
-    Both arms go through ``_carries``, so an argument is held to the same
-    boundary rule as an identifier: with plain containment, "scope='snapshots'"
-    satisfies "scope='snapshot'" while naming a different value.
-
-    Hardcoded on-screen names are left to ``_localised_hardcoded_name``, which
-    applies the pipeline's own rule to them. Excluding them here is structural
-    rather than load-bearing: every name it lists is several words long and
-    none tokenises as a literal, so the filter removes nothing today. It is
-    what keeps a name that later gains an extractable shape from reporting
-    twice under two different descriptions.
-    """
-    protected = _untranslatable_names()
-    return sorted(
-        {
-            stripped
-            for literal in _LITERAL_RE.findall(english)
-            if (stripped := _without_sentence_punctuation(literal))
-            and stripped not in protected
-            and stripped not in _PROSE_ABBREVIATIONS
-            and not _carries(stripped, translated)
-        }
-        | {
-            assignment
-            for assignment in _QUOTED_ASSIGNMENT_RE.findall(english)
-            if not _carries(assignment, translated)
-        }
-    )
-
-
-def _invented_files(english: str, translated: str) -> list[str]:
-    """Config files the translation names and its English does not.
-
-    Literal parity is otherwise one-directional: it asks what the English has
-    that the translation lost, so a translation naming a file of its own
-    passes. That is how three catalogs kept describing ``packages/*.yaml`` and
-    ``themes/*.yaml`` for a tool whose English stopped naming either -- the
-    English literal they *did* carry was still there, and nothing looked at
-    the rest.
-
-    Only files, deliberately, and the measurement is why. Asking the same
-    question of every literal shape reports 62 pairs on the shipped catalogs
-    and 58 of them are correct translations: German, Dutch and Swedish
-    compound with a slash ("Schreib-/Schreib-Tools"), which the path arm reads
-    as a literal the English never had, and Swedish abbreviates "till exempel"
-    as "t.ex". A file name has no such prose shape -- a catalog that names one
-    is naming a file on the reader's disk -- and restricted to that, the same
-    question reports six occurrences across the nine catalogs, all three of
-    them real.
-
-    A dropped duplicate is the other half of the number arm's multiset rule
-    and is not implemented here: no shipped English string names the same file
-    twice, so the arm would be structurally empty rather than verified, and
-    the reverse direction is what the live corpus actually needed.
-    """
-    return sorted(
-        {
-            stripped
-            for literal in _FILE_RE.findall(translated)
-            if (stripped := _without_sentence_punctuation(literal))
-            and not _carries(stripped, english)
-        }
-    )
-
-
-def _localised_hardcoded_name(english: str, translated: str) -> list[str]:
-    """The on-screen name this translation localised away, if any.
-
-    The pipeline already refuses engine output that translates one of these,
-    but that gate only ever runs while *accepting* new output. A catalog edited
-    by hand, or one pinned before the name existed, is never re-read by it:
-    the baseline still matches, so the sync plans no rewrite, and a reader is
-    left hunting for an entry title the integration does not display. Asking
-    the same question of the pinned pairs costs one call and needs no second
-    rule -- ``translate_locales`` owns it, and it is called rather than copied
-    so the two cannot drift apart.
-    """
-    import translate_locales
-
-    dropped = translate_locales._untranslatable_name_dropped(english, translated)
-    return [dropped] if dropped is not None else []
 
 
 @cache
@@ -1728,7 +1315,7 @@ def test_translations_keep_english_numbers_and_identifiers(locale: str) -> None:
     a maintainer decision rather than something to infer here.
 
     A tool string has more than one English rendering, and the check compares
-    against the one the sync sends rather than whichever one matches: eleven of
+    against the one the sync sends rather than whichever one matches: seven of
     the 176 tool keys differ between renderings, and accepting any of them
     would let a fault in the real source hide behind another. Which one that is
     needs no guessing -- ``_plan_settings`` builds its ``tool_texts`` from
@@ -1875,7 +1462,12 @@ def test_literal_extraction_ignores_sentence_punctuation(
         # rows exist: the corpus cannot tell "found nothing" from "not run".
         ("numbers lost", "keeps 30 entries", "behaelt Eintraege", "30"),
         ("numbers added", "keeps entries", "behaelt 30 Eintraege", "30"),
-        ("lost literals", "edit configuration.yaml", "bearbeite die Datei", "configuration.yaml"),
+        (
+            "lost literals",
+            "edit configuration.yaml",
+            "bearbeite die Datei",
+            "configuration.yaml",
+        ),
         ("wrong unit", "limit is 1-256 MB", "Grenze ist 1-256 GB", "256 MB"),
         ("reversed pair", "Range 1-600.", "Bereich 600-1.", "1-600"),
         (
@@ -1966,7 +1558,8 @@ def test_a_number_keeps_its_groups_unless_they_are_thousands(
 
     Every reversal case routes through here and passes under the obvious
     wrong implementation -- join the groups and compare the digits -- which
-    makes 4.5 equal to 45 and the component version 1.2.4 equal to 12.4.
+    makes 4.5 equal to 45 and the minimum component version 1.2.4 equal to
+    12.4.
     Nothing exercised it directly, so that mutation left every shipped pair
     and all three reversal cases green.
     """
